streamlit/requirements.txt
argparse
streamlit
Pillow